Fields will not read what it writes
When to_representation is run on a geo object json.load adds unicode strings so the result is something like:

```
{u'type': u'Point', u'coordinates': [39.921092168090915, 28.04062843322754]}
```

this is the string that rest_framework puts into the web form for any PUTs to edit the field. Therefore the PUT will send the string:

```
"{u'type': u'Point', u'coordinates': [39.921092168090915, 28.04062843322754]}"
```

which will not work with GEOSGeometry as it does not expect the u symbols. The way around this is to eval the string as a python dictionary and then dump the json. This creates the string:

```
'{"type": "Point", "coordinates": [39.921092168090915, 28.04062843322754]}'
```

which will then work.