Added possibility to override GeoJSON properties
By default GeoFeatureModelSerializer will use the attributes of a
model instance (except id, geometry field and bbox field) to fill
the "properties" member of the resulting GeoJSON object.

This commit separates this functionality in a few methods,
which can be overriden by subclasses of GeoFeatureModelSerialzier.
This way users will be able to override the default behaviour in
order to supply their own source for the properties member.