Implement custom feature properties source
By default, when converting a django model instance with geometry
to GeoJSON, the "properties" field will be filled with the fields
defined in the corresponding serializer.

In this commit we separate the code a bit in individual methods,
which can be overriden by any subclass. In this way the user is
able to define their own source for the properties field of the
GeoJSON output.