[fix] Deserialize id_field #272
When serializing a model, the value of its `id_field` is written in the `id` attribute
of the GeoJSON feature.

This behaviour is however not mirrored by the deserialization logic: deserializing a
model through GeoFeatureModelSerializer ignores the `id` field set in the given GeoJSON
feature.

This commit solves this issue by mirroring the serialization logic:
- When deserializing a model, the `id` attribute of the GeoJSON feature is written as the
`id_field` of the django model.

Fixes #272